Add datMap/datUnmap
Not supported by HDF5. Presumably because of things like
compression filters and data type conversion.

datMap must therefore mmap and anonymous memory area to
receive the data and then datUnmap must copy the data back
to the underlying dataset.

Will require that datMap and datUnmap lose the const in
the API as HDSLoc will need to be updated. Also need datAnnul
to automatically do the copy.

Unsolved problems (apart from no mmap):

What happens if the primitive is mapped twice? With different
locators? What happens in HDS?

What happens if you call datPut/datGet whilst the locator
is memory mapped?

Do you have to track all the mapped locators and put
some code in to try to unmap them all on program exit?

If you annul the root locator should you go and annul
all the child locators so that everything is synced
before the file is closed?