Use routine for copying dimensions
HDS and HDF5 use different (in theory) data types
for specifying dimensions. Move all the code for copying
from one type to another into two separate routines (one
for import and one for export).

It may well be that we have to transpose the dimensions
to handle the C order used by HDF5 and match the Fortran
order (HDF5 already transposes in its Fortran interface).