Modify Fortran locator I/O
No longer clone the locator as this doesn't work for file mapping
which does not copy to the clone. We now simply treat the Fortran
locator as if it is the pointer and is just extracted and passed
around.

dat1_import_floc is now removed and datImportFloc should be
used everywhere.