No longer create resizable datasets
Resizable datasets can not be memory-mapped. Memory mapping
is important so disable resizing and force datAlter to
resize by creating a new dataset, copying the data in, deleting
the original and then renaming the temporary. The hope is that
datAlter is called far less often than datMap and we will gain
in speed overall (possibly at the expense of file size but that
is to be determined).