Attempt to mmap() in datMap
If an offset can be retrieved from H5Dget_offset and if the
data type on disk matches the data type being requested, then
it is possible that mmap() could be used.

Currently this seems to be unreliable even when restricted
to memory mapping readonly files so the code is currently
disabled by explicitly setting the try_mmap variable to false
in line 255.

If that line is removed an attempt will be made to memory map
readonly files and this mostly works but ccdexercise gets
confused with the final mosaic and this has not been debugged.