datMap: Use VFD interface to get file descriptor
This only seems to be reliable in READONLY mode but it does mean
the file does not need to be opened separately. Also, this raised
the possibility that a mmapped buffer may or may not be associated
with a file handle that should be closed so we now flag that
state independently in the locator.