hdsFlush: Remove from group membership before annulling
This fixes a problem whereby the walk through the list
of locators was finishing early because datAnnul was calling
back to the group code to remove the locator from the group.
This obviously messed up all the counting and some locators
were leaking.

Now the code removes the locator from the group before calling datAnnul.