Experimental check of file type
check if a file is HDF5 format before attempting to open it. This
may allow us to have a better error message trap in the wrapper
HDS library as it is important to distinguish "error from opening
an HDF5 file" (in which case don't bother trying HDSv4) versus
"this is not an HDF5 file" (so it's probably an HDSv4 file).

Unfortunately if you have opened a new file for write but not closed
it before opening for update, this test will fail. This probably
means the test is useless but we will see what happens.