hdsFree: Flush buffers to disk
hdsFree is meant to be the complementary routine to hdsLock.
It seems though that despite no Starlink code ever calling hdsLock
it is critical that hdsFree forces a flush of buffers to disk. This
seems to be an undocumented feature of hdsFree in the v4 API.

Without this fix ICL can not operate properly and ORAC-DR fails
to start. From ICL you can see this with the CCDEXERCISE test
in CCDPACK which fails with the error:

   CCDGENERATE
    ===========

   Generating test data.

!! SUBPAR: Error finding component '67507' in $ADAM_USER/icl.67507.VARIABLES.OBJECT_FILE
!  SUBPAR: Failed to open HDS object associated with parameter FILE
!  SUBPAR: Bad preset value for parameter FILE
FILE - List containing object positions etc. /@$ADAM_USER/icl.66809.VARIABLES.OBJECT_FILE/ > !!

CCDEXERCISE runs fine with this patch.