datAnnul: Split in two
datAnnul still behaves as before in that it frees file resources
and frees the struct memory, returning a NULL pointer. The
file resource freeing now occurs in dat1Annul so that automated
freeing of locators when the last primary locator is annulled
can happen without the possibility of triggering a double
free if the caller code then explicitly annuls the secondary
locator.

This will result in some small leaking of C struct memory
and probably needs a routine to allow all these orphaned
structs to be freed properly