Tweak the calculation of primary/secondary locators
HDS closes a file and annuls all locators when the last primary
locator associated with a file is freed. This was implemented
in v5 by using a per-file_id accounting of primary locators. This
does not work as expected though because HDF5 issues a new file_id
even when the same file descriptor is being used. The count for
primary should include all file_ids associated with a single
file.

The code now retrieves the underlying file descriptor and
frees all file_ids-associated locators when the total primary
count for the file descriptor hits zero.

Without this ndfcopy of a SMURF makemap file simply did
not work as the locators were annulled too early.