Rewrite handling of primitive STATE
Previously an attribute was used which tracked whether the primitive
had been written to or not. Now the attribute has been removed and
we call H5Dget_space_status directly. This seems to behave in the
same way although datReset is now implemented by deleting the
entire primitive and re-creating it empty.