datSlice: Handle cases where the supplied locator is already a slice
If the supplied locator is already a slice, the supplied lower and upper
bounds will refer to the grid system of the slice, not of the associated
dataset. So modify the supplied lower and upper bounds so that they refer
to the grid space of the dataset by adding on the lower bounds of the
existing slice (if any).