Report informative error if input locator has been annulled
The container file is closed when the last primary locatyor is annulled,
and any remaining secondary locators are annulled automatically. Any
future use of an annulled secondary locator previously caused spurious
unrelated error messages to be reported (for instance, I got "Unexpected
locator with no dataspace"). The changes in this commit cause a more
informative error message to be issued.

Each public function now checks the validity of the input locators and
reports an error if they are not valid (e.g. if they have been annulled).