Attempt to link a locator to a group even if no rw lock has been obtained
Since the group name is stored inside the locator, assigning a locator to
a group requires a change to the locator structure, which means that the
calling thread needs a read-write lock on the locator to prevent possible
simultaneous access from another thread. However, if the locator is for
an HDS object that has been opened read-only, the current thread will
only have a read lock, and not a read-write lock. This commit attemts to
get round this by promoting the read lock temporarily to a read-write
lock before assigning a locator to a group. This can only be done if the
current thread has the only read-lock (i.e. no other threads also have
read locks).