Change to use signed 64 bit ints for HDS dimensions
This needs to be done since the main point of the move to HDS-V5 is to
allow starlink apps to process arrays with more than 2^32 pixels. Since
the ARY library standardly vectorises each N-D arrray and then accesses
it as a 1D array, individual axis dimensions can also be greater than
2^32. They need to be signed since hdsdim will also be used for ARY/NDF
bounds, which can be negative.