Provide a macro for the HDS data type to use with dimension values
The macro HDS_DIM_type is "K" for 64 bit dimensions and "I" for 32 bit
dimensions.

Another reason for using signed values for hdsdim is that we
do not currently provide a "UK" (unsigned 64 bit int) data type in HDS,
and so could not store hdsdim values within an HDS data file if hdsdim
was unsigned (as is needed by ARY/NDF).