Allow f77 apps to bypass internal HDS lock checks
This is because multi-threaded C apps that use F77 code (via a suitabel C
wrapper) serialise the execution of the F77 code using a mutex (usually
within the wrapper layer), and so HDS functions called within the F77
code (via a suitable wrapper) should not perform extra serialisation
checks.

The new datNolock function can be used to indicate that locking checks
should not be performed on a specified object.