Add support for blank HDS structure type names
HDS V5 allows structures to have blank names (as used within ARY etc), so
HDS-V5 should also allow the same thing. Previously, a blank type
name caused an error when it was stored as an HDF5 attribute, because
the blank name was normalised to a null (zero-length) string within
dau1CheckType.c. Now it is normalised to a single space.