hds-v5: New tuning parameter HDS_LOCKCHECK
If non-zero, an error is reported if a V5 HDS object has not been locked
appropriately by the current thread before use. If zero, no such check
is performed. This is provided so that old software which pre-dates the
datLock function (and which presumably is explicitly designed to avoid
conflicting simultaneous access to an HDS object from separate threads)
can be used without change. Use of this tuning parameter should be seen
as a temporary stop-gap measure until such time as the software can be
changed to use the datLock function correctly.