Avoid use of starlink-specific HDF5 code when reopening in read-write mode
The starlink-specific facility within HDF5 for re-opening read-only files
with read-write access seems to have broken in HDF5 V1.10.4. So instead,
this commit changes HDS-V5 so that, when asked to open a file in
WRITE/UPDATE mode that has already been opened in READ mode, it first
records all active locators associated with the file, then closes the
file, then re-opens it with the new access mode, and then modifies all
the active locators to refer to the newly opened file.