Correct bug that caused hdsErase to have no effect
Store the deferred file deletion flag in the handle rather than the
locator. The old system did not work if there were other locators
refering to the same file, because the flag got deleted along with the
locator and so had no effect when later locators referring to the same
file were annulled. All locators for a single file use the same Handle
object, so putting the deletion flag in the Handle allows the flag to be
honoured no matter which locator ends up being annilled last.