datUnlock: Do not report an error if the object is already unlocked
Changed so that no error is reported if the object is already unlocked on
entry. This is because there can be multiple locators for the same object
- unlocking an object using one such locator should not cause an error if
a later attempt is made to unlock it using one of the othjer locators.
For instance, in ARY the "loc" and "dloc" locators may sometimes refer to
the same HDS object. Before this change aryUnlock reported an error in
such cases because unlocking "loc" also implicitly caused "dloc" to be
unlocked, so that the subsequent attempt to unlock "dloc" explicitly
caused an error to be reported. There is no HDS function to tell if two
locators refer to the same object, so the alternative approach ofnot
attempting to unlock "dloc" if it refers to the same object as "loc",
cannot be used.