datMsg: Indicate token is undefined if an error occurs.
Previously, if an error occurred getting the path for a locator the
message token would be left with the value it had on entry to datMsg.
This causes misleading error messages. So now, if an error occurs getting
the path for the locator, set the token value to "^<token name>", which
is the standard way to indicate that a message token has no defined
value.