Ensure all HDF5 ids get closed when last primary locator is annulled
If two primary locators exist for the same file but use different HDF5
file_ids, annulling the first one will not cause the file to be closed,
so the HDF5 file_id is left open. When the second one is annulled, the
file is closed, but previously, this only resulted in the file_id
associated with the second locator being closed. The file_id associated
with the first locator was till open, causing HDF5 to keep the file open.
This caused all sorts of spurious behaviour when used from within a
monolith as files could be left open between invocations of different
commands.

This commit ensures that annulling the final primary locator causes all
remaining HDF5 identifiers for the same file to be closed.